requirements.txt
Flask>=2.2.0
google-cloud-vision
gunicorn
